rm useless component, rm default filters
